Store Logic

+++ Scripts/Netcode/ENetClient.cs
@@ -326,9 +326,9 @@ namespace KRU.Networking
             }
         }
 
-        public static void PurchaseItem(int itemId)
+        public static void PurchaseItem(uint itemId)
         {
-            var data = new WPacketPurchaseItem { StructureID = (ushort)itemId };
+            var data = new WPacketPurchaseItem { StructureID = (uint)itemId };
             var clientPacket = new ClientPacket((byte)ClientPacketOpcode.PurchaseItem, data);
 
             Outgoing.Enqueue(clientPacket);
